+++ frb/data/Galaxies/build_hosts.py
@@ -70,10 +70,13 @@ def host_121102():
 
     # Derived quantities
     host121102.derived['M_r'] = -17.0  # AB; Tendulkar+17
+    host121102.derived['M_r_err'] = 0.2  # Estimated by JXP
     host121102.derived['SFR_nebular'] = 0.23  # MSun/yr; Tendulkar+17
     host121102.derived['Mstar'] = 5.5e7  # Msun; Tendulkar+17
     host121102.derived['Mstar_err'] = 1.5e7  # Msun; Tendulkar+17
-    
+    host121102.derived['Z_spec'] = -0.16  # Tendulkar+17 on a scale with Solar O/H = 8.86
+    host121102.derived['Z_spec_err'] = -999.  # Tendulkar+17
+
     # Vet
     assert host121102.vet_all()