+++ CHANGES.rst
@@ -3,4 +3,10 @@
 
 - Eliminate 2.7
 - Start igm.py module
+- Add RM and associated methods
+- Add DM_ISM
+- Add FGGalaxy object
+- Update halo mass calculations for the new Aemulus HMF
+- Add Virgo halo
+- Simple Milky Way methods for DM